(DAY1 - LOGIC) : pulling in data from the dummy-data file and refactored some file structure. also started a PostContainer component
