+++ instagram/src/components/PostContainer/PostContainer.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const PostContainer = props => {
+  return(
+    <div className="post-container-wrapper">
+      <h1>This is the post container working title</h1>
+    </div>
+  );
+};
